requirements.txt
fastapi
uvicorn
python-pptx
python-multipart
requests
mcp[cli]>=1.10.1
# wheel（如需本地打包时用，可放最后或注释掉）
python-docx
openpyxl
Pillow
requests
# PDF解析相关依赖
PyPDF2
pdfplumber
